requirements.txt
fastapi
uvicorn[standard]
chromadb
requests
beautifulsoup4
python-dotenv
sentence-transformers
openai>=1.40.0
